refactor: split main.rs into cli, commands, send, validate, generate modules
Split logfwd main.rs (3,384 lines) into 6 files:
- main.rs (150 lines) — entry point only
- cli.rs — clap definitions and CLI types
- commands.rs — command handlers, wizard, config discovery
- send.rs — ff send config building
- validate.rs — config and pipeline validation
- generate.rs — synthetic log generation

All 61 unit + 18 integration tests pass.
